+++ src/headless_re_mcp/backends/r2/client.py
@@ -29,6 +29,7 @@ _ALLOWED = frozenset(
         "ie",
         "aflj",
         "izj",
+        "izzj",
         "iij",
         "iEj",
         "pdj",